Re-estructures pakages. Changed name from eda to dsa. Created a pakage for each data structure. Implemented Map, MapEntry,HashTable, MapFactory and half MapTest.
