+++ java/src/dsa/FIFO/model/Queue.java
@@ -1,0 +1,9 @@
+package dsa.FIFO.model;
+
+public interface Queue<T> {
+    void add(T e);
+    T poll();
+    T peek();
+    boolean isEmpty();
+    String toString();
+}
