+++ java/src/dsa/list/model/List.java
@@ -1,0 +1,7 @@
+package dsa.list.model;
+
+/**
+ * Created by JoseLlorens on 20/04/2016.
+ */
+public interface List {
+}
