+++ java/src/dsa/sorting/SortingFactory.java
@@ -1,0 +1,12 @@
+package dsa.sorting;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class SortingFactory implements Sorting.Factory {
+
+    @Override
+    public List<Sorting> create() {
+        return Arrays.asList(new InsertionSort(), new MergeSort(), new QuickSort(), new SelectionSort(), new BubbleSort());
+    }
+}
