+++ java/src/dsa/stack/Stack.java
@@ -1,0 +1,7 @@
+package dsa.stack;
+
+/**
+ * Created by JoseLlorens on 20/04/2016.
+ */
+public interface Stack {
+}
