--- java/src/eda/FIFO/model/Queue.java
@@ -1,9 +1,0 @@
-package eda.FIFO.model;
-
-public interface Queue<T> {
-    void add(T e);
-    T poll();
-    T peek();
-    boolean isEmpty();
-    String toString();
-}
