--- java/src/eda/data_structures/models/IPList.java
@@ -1,7 +1,0 @@
-package eda.data_structures.models;
-
-/**
- * Created by JoseLlorens on 20/04/2016.
- */
-public interface IPList {
-}
