--- java/src/eda/data_structures/models/Queue.java
@@ -1,23 +1,0 @@
-package eda.data_structures.models;
-
-/**
- * Created by JoseLlorens on 20/04/2016.
- */
-public interface Queue<T> {
-
-    void add(T x);
-
-    T element();
-
-    T peek();
-
-    T poll();
-
-    T remove();
-
-    void removeAll();
-
-    interface Factory{
-        java.util.List<String> create();
-    }
-}
